Refactor chat interface and command structure
- Replaced `LiveChatJsonInterface` with `LiveChatJsonCommand` for better command handling.
- Introduced `FormattedJsonCommand` to manage formatted JSON data operations.
- Updated related tests to utilize the new command structure.
- Removed deprecated interface modules from the codebase.
