+++ youtube/src/interface.rs
@@ -2,6 +2,4 @@
 //! The interface layer is the layer that interacts with the outside world.
 
 pub mod cli;
-pub mod formatted_json_interface;
-pub mod live_chat_json_interface;
 pub mod presenter;
