+++ youtube/src/interface/cli.rs
@@ -1,3 +1,7 @@
-mod watch_chat_command;
+pub mod formatted_json_command;
+pub mod live_chat_json_command;
+pub mod watch_chat_command;
 
+pub use formatted_json_command::FormattedJsonCommand;
+pub use live_chat_json_command::LiveChatJsonCommand;
 pub use watch_chat_command::WatchChatCommand;
