+++ youtube/src/interface/cli/formatted_json_command.rs
@@ -1,15 +1,14 @@
 use crate::application::chat_service::ChatConvertService;
 use crate::infrastructure::io::chat_service_repository::IoChatServiceRepository;
-use chrono::prelude::*;
 use std::path::{Path, PathBuf};
 
 /// This service provides an interface for managing and retrieving live chat JSON data from files.
-pub struct LiveChatJsonInterface<'a, T> {
+pub struct FormattedJsonCommand<'a, T> {
     source: &'a T,
 }
 
-impl<'a, T> LiveChatJsonInterface<'a, T> {
-    /// Create a new LiveChatJsonService instance.
+impl<'a, T> FormattedJsonCommand<'a, T> {
+    /// Create a new FormatedJsonService instance.
     ///
     /// # Arguments
     /// - `inner`: Source file path.
@@ -19,7 +18,7 @@ impl<'a, T> LiveChatJsonInterface<'a, T> {
 }
 
 /// This implementation is for the PathBuf type.
-impl LiveChatJsonInterface<'_, PathBuf> {
+impl FormattedJsonCommand<'_, PathBuf> {
     /// Generate simple chat CSV data from live chat JSON data.
     ///
     /// # Arguments
@@ -35,7 +34,7 @@ impl LiveChatJsonInterface<'_, PathBuf> {
 
         let service = ChatConvertService::new(repositories);
 
-        service.convert_from_lines().await
+        service.convert_from_chunk().await
     }
 
     /// Generate simple chat CSV data from live chat JSON data.
@@ -46,7 +45,6 @@ impl LiveChatJsonInterface<'_, PathBuf> {
         let source_path = self.source.clone();
         let mut target_path = source_path.clone();
         target_path.set_extension(file_type);
-        let target_path = target_path;
 
         let repositories = vec![IoChatServiceRepository::file_to_file(
             source_path,
@@ -54,12 +52,29 @@ impl LiveChatJsonInterface<'_, PathBuf> {
         )?];
 
         let service = ChatConvertService::new(repositories);
+        service.convert_from_chunk().await
+    }
 
-        service.convert_from_lines().await
+    /// Generate simple chat CSV data from live chat JSON data.
+    ///
+    /// # Returns
+    /// - `anyhow::Result<String>`: Result of the conversion.
+    pub async fn generate_string(&self) -> anyhow::Result<String> {
+        let source_path = self.source.clone();
+
+        let repositories = vec![IoChatServiceRepository::file_to_in_memory(source_path)?];
+
+        let service = ChatConvertService::new(repositories);
+        service.convert_from_chunk().await?;
+
+        let repositories = service.move_chat_service_repository();
+
+        let data_str = repositories.first().unwrap().to_in_memory_data()?;
+        Ok(data_str)
     }
 }
 
-impl LiveChatJsonInterface<'_, Vec<PathBuf>> {
+impl FormattedJsonCommand<'_, Vec<PathBuf>> {
     /// Generate simple chat CSV data from live chat JSON data.
     pub async fn generate_files_with_csv(&self) -> anyhow::Result<()> {
         let source_paths = self.source.clone();
@@ -80,43 +95,17 @@ impl LiveChatJsonInterface<'_, Vec<PathBuf>> {
 
         let service = ChatConvertService::new(repositories);
 
-        service.convert_from_lines().await
-    }
-
-    /// Generate simple chat CSV data from live chat JSON data.
-    pub async fn generate_files_with_timestamped_name(&self) -> anyhow::Result<()> {
-        let source_paths = self.source.clone();
-
-        let results = source_paths
-            .into_iter()
-            .map(|source_path| {
-                let created_at = source_path.metadata()?.created()?;
-                let created_at: DateTime<Local> = created_at.into();
-
-                let target_path = source_path
-                    .with_file_name(format!("{}.csv", created_at.format("%Y-%m-%dT%H-%M-%S%z")));
-
-                let rp = IoChatServiceRepository::file_to_file(source_path, target_path)?;
-
-                Ok(rp)
-            })
-            .collect::<Vec<_>>();
-
-        let repositories = rust_support::anyhow::collect_results(results)?;
-
-        let service = ChatConvertService::new(repositories);
-
-        service.convert_from_lines().await
+        service.convert_from_chunk().await
     }
 }
 
 /// This implementation is for the String type.
-impl LiveChatJsonInterface<'_, String> {
+impl FormattedJsonCommand<'_, String> {
     /// Generate simple chat CSV data from live chat JSON data.
     ///
     /// # Arguments
-    /// - `target_path`: The path to save the converted data.
-    pub async fn generate_file_with_string(&self, target_path: &Path) -> anyhow::Result<()> {
+    /// - `to_path`: The path to save the converted data.
+    pub async fn generate_file_with_path(&self, target_path: &Path) -> anyhow::Result<()> {
         let source_string = self.source.clone();
         let target_path = target_path.to_path_buf();
 
@@ -127,6 +116,26 @@ impl LiveChatJsonInterface<'_, String> {
 
         let service = ChatConvertService::new(repositories);
 
-        service.convert_from_lines().await
+        service.convert_from_chunk().await
+    }
+
+    /// Generate simple chat CSV data from live chat JSON data.
+    ///
+    /// # Returns
+    /// - `anyhow::Result<String>`: Result of the conversion.
+    pub async fn generate_string(&self) -> anyhow::Result<String> {
+        let source_string = self.source.clone();
+
+        let repositories = vec![IoChatServiceRepository::in_memory_to_in_memory(
+            source_string,
+        )?];
+
+        let service = ChatConvertService::new(repositories);
+        service.convert_from_chunk().await?;
+
+        let repositories = service.move_chat_service_repository();
+
+        let data_str = repositories.first().unwrap().to_in_memory_data()?;
+        Ok(data_str)
     }
 }
