+++ youtube/src/lib.rs
@@ -6,8 +6,8 @@ mod interface;
 pub mod prelude {
     //! This module contains the prelude for the library.
 
+    pub use crate::interface::cli::FormattedJsonCommand;
+    pub use crate::interface::cli::LiveChatJsonCommand;
     pub use crate::interface::cli::WatchChatCommand;
-    pub use crate::interface::formatted_json_interface::*;
-    pub use crate::interface::live_chat_json_interface::*;
     pub use crate::interface::presenter::StdoutPresenter;
 }
